chore(suite): move thp modals into connection/thp
