+++ packages/suite/src/components/connection/thp/ThpAutoconnectionModal.tsx
@@ -1,27 +1,24 @@
 import { TrezorDevice } from '@suite-common/suite-types';
 import { thpActions } from '@suite-common/thp';
-import TrezorConnect from '@trezor/connect';
 
-import { useDispatch } from 'src/hooks/suite';
+import { useDispatch } from '../../../hooks/suite';
+import { ConfirmActionModal } from '../../suite/modals/ReduxModal/DeviceContextModal/ConfirmActionModal';
 
-import { ConfirmActionModal } from '../suite/modals/ReduxModal/DeviceContextModal/ConfirmActionModal';
-
-type ThpConnectionModalProps = {
+type ThpAutoconnectionModalProps = {
     device: TrezorDevice;
 };
 
-export const ThpConnectionModal = ({ device }: ThpConnectionModalProps) => {
+export const ThpAutoconnectionModal = ({ device }: ThpAutoconnectionModalProps) => {
     const dispatch = useDispatch();
 
     const onCancel = () => {
-        TrezorConnect.cancel();
         dispatch(thpActions.resetThpFlow());
     };
 
     return (
         <ConfirmActionModal
             device={device}
-            title="TR_THP_SECURELY_CONNECT_WITH_TREZOR"
+            title="TR_THP_SECURELY_AUTOCONNECT_WITH_TREZOR"
             onCancel={onCancel}
             enableBackdropClick={false}
         />
