+++ packages/suite/src/components/connection/thp/ThpConnectionModal.tsx
@@ -1,24 +1,27 @@
 import { TrezorDevice } from '@suite-common/suite-types';
 import { thpActions } from '@suite-common/thp';
+import TrezorConnect from '@trezor/connect';
 
-import { useDispatch } from '../../hooks/suite';
-import { ConfirmActionModal } from '../suite/modals/ReduxModal/DeviceContextModal/ConfirmActionModal';
+import { useDispatch } from 'src/hooks/suite';
 
-type ThpAutoconnectionModalProps = {
+import { ConfirmActionModal } from '../../suite/modals/ReduxModal/DeviceContextModal/ConfirmActionModal';
+
+type ThpConnectionModalProps = {
     device: TrezorDevice;
 };
 
-export const ThpAutoconnectionModal = ({ device }: ThpAutoconnectionModalProps) => {
+export const ThpConnectionModal = ({ device }: ThpConnectionModalProps) => {
     const dispatch = useDispatch();
 
     const onCancel = () => {
+        TrezorConnect.cancel();
         dispatch(thpActions.resetThpFlow());
     };
 
     return (
         <ConfirmActionModal
             device={device}
-            title="TR_THP_SECURELY_AUTOCONNECT_WITH_TREZOR"
+            title="TR_THP_SECURELY_CONNECT_WITH_TREZOR"
             onCancel={onCancel}
             enableBackdropClick={false}
         />
