+++ packages/suite/src/components/connection/thp/ThpPairingFailedForFirmwareInstallation.tsx
@@ -1,5 +1,5 @@
 import { ThpPairingCodeEntry } from './ThpPairingCodeEntry';
-import { useSelector } from '../../hooks/suite';
+import { useSelector } from '../../../hooks/suite';
 
 export const ThpPairingFailedForFirmwareInstallation = () => {
     const lastThpCode = useSelector(state => state.thp.lastThpCode);
