+++ packages/suite/src/views/firmware/Steps/StepThpPairing.tsx
@@ -3,8 +3,8 @@ import { ReactNode } from 'react';
 import { Card, Column, Modal, Text } from '@trezor/components';
 import { spacings } from '@trezor/theme';
 
+import { ThpPairingCodeEntry } from '../../../components/connection/thp/ThpPairingCodeEntry';
 import { Translation } from '../../../components/suite';
-import { ThpPairingCodeEntry } from '../../../components/thp/ThpPairingCodeEntry';
 
 type StepThpPairingProps = {
     modalHeading: ReactNode;
